bridge_agent/pqc_signing.py
import oqs # liboqs-python must be installed

class PQCSigner:
    def __init__(self):
        self.alg = "ML-DSA-65" # NIST FIPS 204 Standard

    def create_pqc_auth(self, message_bytes):
        """
        Signs the release message using a Quantum-Resistant key.
        """
        with oqs.Signature(self.alg) as signer:
            public_key = signer.generate_keypair()
            private_key = signer.export_secret_key()
            
            signature = signer.sign(message_bytes)
            return public_key, signature

# Foundation Test logic
if __name__ == "__main__":
    pqc = PQCSigner()
    msg = b"RELEASE_SOL_VAULT_001"
    pk, sig = pqc.create_pqc_auth(msg)
    print(f"PQC Public Key Generated: {pk.hex()[:20]}...")
    print(f"Quantum-Resistant Signature: {sig.hex()[:20]}...")